+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 * Youtube controls
 * Default controls
 * Temporary fix for https://github.com/flutter/flutter/issues/50686 
+
+## [0.0.3] - 
+* Bug fix default controls set by default
+* Default error widget - Retry button that re-initiates the player